requirements.txt
imageio==2.31.3
matplotlib==3.7.2
mmcv==2.0.1
mmdet==3.1.0
mmengine==0.8.4
numpy==1.25.2
opencv_python==4.8.0.76
Pillow==9.4.0
Pillow==10.1.0
pycocotools==2.0.7
torch==2.0.1
tqdm==4.65.2
